I gave up on giving up.
So, I'm now working on getting a list of authors and their works, like this.
Catullus: Carmina
I'm using a pyhton script to go through the latin library site and do that,
but I'm having a bit of a hard time. It's catching \n and navigation text
and stuff, and that's no good, and it keeps getting caught on some weird
unicode thing, apparently. I dunno. You can run it and see the error. I have
no idea what's wrong. But I do want to finish this, someday...