Started on adding pre-pasted texts.
I need to get the texts from the website, but without the newline characters,
so I'm working on a python script to give them to me as one huge line with
"\n" in it so I can just paste it into texts.js. I think what I should do
is paste them manually into a file, tell python to replace "\n" with "\\n",
and then paste that into javascript. That should work. I think.