Redesigned things a little bit.
I made the list of works dynamically generated so that it's easier
to add stuff. You just put in in the javascript and it's done.
No more playing with HTML. I also made it categorize the works
by author so I can have more and still use up little space.